Use static dispatch in various methods
The potential penalty of dynamic dispatch is unknown in these cases,
and currently there is no real benefit to using dynamic dispatch.
We might revisit these cases in the future, but if we do we should do
so with benchmarks to determine whether or not the change leads to
significant performance differences.